Small code tidy to more directly state the logic.
Use 'all' instead of a loop to check if the displays have changed.

Reduced the amount of content inside the tuple in 'getGeometry' to make it easier to read. It also happens to save a few calls to `screen.geometry`.

Replace a small block in 'restartClocks' with an already existing function that does the same thing.