Update german language to v3.5.0
Last translation I might do for a while as I dont really use the software anymore since windows 11 now has a built-in clock on the second monitor. That was really the only reason I used this software in the first place. I really think that the other features are cool as well, but I just don't need them. Still an awesome software, keep up the great work! I may check on the translations here occasionally if I have time and feel like translating. Anyways, here are the german translations for v3.5.0.